+++ stats/effects/st_avianFallResist/st_avianFallResist.lua
@@ -1,3 +1,9 @@
 function init()
-  effect.addStatModifierGroup({{stat = "fallDamageMultiplier", effectiveMultiplier = 0.25}})
+
+  -- Get configuration for fall damage
+  multFalling = config.getParameter("multFalling", 1)
+
+  -- Set bonus based on config values
+  effect.addStatModifierGroup( { {stat = "fallDamageMultiplier", effectiveMultiplier = multFalling } } )
+
 end